add readme and integration test

+++ src/location.rs
@@ -3,15 +3,15 @@ use std::{fs, io};
 use std::ffi::{OsString};
 use std::fs::{DirEntry};
 use std::path::{Path, PathBuf};
-use std::time::UNIX_EPOCH;
 use cached::cached_result;
 use cached::instant::SystemTime;
 use crate::config::Location;
+use crate::utils::time_as_num;
 
 use super::config;
 use super::utils;
 
-#[derive(Serialize, Deserialize, Debug, Clone)]
+#[derive(Serialize, Deserialize, Debug, Clone, PartialEq)]
 pub struct LocationEntry {
     pub packet: String,
     pub time: f64,
@@ -86,7 +86,7 @@ pub fn mark_packet_known(packet_id: &str, location_id: &str, hash: &str, time: S
     let schema_version = config::read_config(root)?.schema_version;
     let entry = LocationEntry {
         packet: String::from(packet_id),
-        time: time.duration_since(UNIX_EPOCH).unwrap().as_millis() as f64,
+        time: time_as_num(time),
         hash: String::from(hash),
         schema_version,
     };
@@ -98,14 +98,18 @@ pub fn mark_packet_known(packet_id: &str, location_id: &str, hash: &str, time: S
 
     fs::create_dir_all(&location_path)?;
     let path = location_path.join(packet_id);
-    fs::File::create(&path)?;
-    let json = serde_json::to_string(&entry)?;
-    fs::write(path, json)?;
+    if !path.exists() {
+        fs::File::create(&path)?;
+        let json = serde_json::to_string(&entry)?;
+        fs::write(path, json)?;
+    }
     Ok(())
 }
 
 #[cfg(test)]
 mod tests {
+    use std::time::{Duration, SystemTime};
+    use crate::test_utils::tests::get_temp_outpack_root;
     use super::*;
 
     #[test]
@@ -114,5 +118,49 @@ mod tests {
         assert_eq!(entries[0].packet, "20170818-164847-7574883b");
         assert_eq!(entries[1].packet, "20170818-164830-33e0ab01");
         assert_eq!(entries[2].packet, "20180818-164043-7cdcde4b");
+    }
+
+    #[test]
+    fn can_mark_known() {
+        let root = get_temp_outpack_root();
+        let loc_a = Path::new(root.as_path()).join(".outpack/location/ae7a7bcb");
+        let entries_a = read_location(loc_a).unwrap();
+        let entry_a = entries_a.first().unwrap();
+        let loc_b = Path::new(root.as_path()).join(".outpack/location/be7a7bcb");
+        let entries_b = read_location(loc_b.clone()).unwrap();
+        assert!(entries_b.iter().find(|e| e.packet == entry_a.packet).is_none());
+        let now = SystemTime::now();
+        mark_packet_known(&entry_a.packet, "be7a7bcb", &entry_a.hash,
+                          SystemTime::now(), root.as_path().to_str().unwrap()).unwrap();
+        let entries_b = read_location(loc_b).unwrap();
+        let res = entries_b.iter().find(|e| e.packet == entry_a.packet).unwrap();
+        assert_eq!(res.time, time_as_num(now));
+        assert_eq!(res.packet, entry_a.packet);
+        assert_eq!(res.hash, entry_a.hash);
+        assert_eq!(res.schema_version, entry_a.schema_version);
+    }
+
+    #[test]
+    fn marking_known_does_not_overwrite() {
+        let root = get_temp_outpack_root();
+        let loc_a = Path::new(root.as_path()).join(".outpack/location/ae7a7bcb");
+        let entries_a = read_location(loc_a).unwrap();
+        let entry_a = entries_a.first().unwrap();
+        let now = SystemTime::now();
+        mark_packet_known(&entry_a.packet, "be7a7bcb", &entry_a.hash,
+                          SystemTime::now(), root.as_path().to_str().unwrap()).unwrap();
+
+        let loc_b = Path::new(root.as_path()).join(".outpack/location/be7a7bcb");
+        let entries_b = read_location(loc_b.clone()).unwrap();
+        let res = entries_b.iter().find(|e| e.packet == entry_a.packet).unwrap();
+        assert_eq!(res.time, time_as_num(now));
+
+        mark_packet_known(&entry_a.packet, "be7a7bcb", &entry_a.hash,
+                          SystemTime::now() + Duration::from_secs(120), root.as_path().to_str().unwrap()).unwrap();
+
+        let entries_b = read_location(loc_b).unwrap();
+        let res = entries_b.iter().find(|e| e.packet == entry_a.packet).unwrap();
+        // time known should still be the time it was first added at
+        assert_eq!(res.time, time_as_num(now));
     }
 }
